FOOD DELIVERY BACKEND SYSTEM/db/db_config.py
import mysql.connector

def get_connection():
    return mysql.connector.connect(
        host = "localhost",
        user = "root",
        password = "vikas@1020",
        database = "food_app"
    )