FOOD DELIVERY BACKEND SYSTEM/requirements.txt
mysql-connector-python
pyjwt
redis
pytest